+++ src/Model/Team.java
@@ -5,6 +5,7 @@ import processing.core.PShape;
 import processing.data.JSONArray;
 import processing.data.JSONObject;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 public class Team {
@@ -15,7 +16,7 @@ public class Team {
     @NotNull
     public String name;
     @NotNull
-    public Integer squadMarketValue;
+    public BigDecimal squadMarketValue;
     @NotNull
     PShape emblem;
 
@@ -23,7 +24,7 @@ public class Team {
         fixtures = getFixtures(team.getJSONObject("_links").getJSONObject("fixtures").getString("href"));
         players = getPlayers(team.getJSONObject("_links").getJSONObject("players").getString("href"));
         name = team.getString("name");
-        squadMarketValue = Util.getInteger(team.getString("squadMarketValue", "").replaceAll("[^\\d]+", ""));
+        squadMarketValue = Util.getBigDecimal(team.getString("squadMarketValue", "").replaceAll("[^\\d]+", ""));
         //emblem = loadShape(teamContent.getString("crestUrl"));
         //emblem = loadShape("https://upload.wikimedia.org/wikipedia/commons/c/c5/Logo_FC_Bayern_M%C3%BCnchen.svg");
     }
